Fix HotkeyControl virtual key display
* A new interop project was setup to provide wrappers for C# projects
  that want to access functionality in the common project.
