+++ src/common/common-md-flag/framework.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#define WIN32_LEAN_AND_MEAN             // Exclude rarely-used stuff from Windows headers
