+++ src/common/keyboard_layout.h
@@ -1,0 +1,20 @@
+#pragma once
+#include <string>
+#include <vector>
+#include <memory>
+#include <Windows.h>
+
+class LayoutMap
+{
+public:
+    LayoutMap();
+    ~LayoutMap();
+    void UpdateLayout();
+    std::wstring GetKeyName(DWORD key);
+    std::vector<DWORD> GetKeyCodeList(const bool isShortcut = false);
+    std::vector<std::wstring> GetKeyNameList(const bool isShortcut = false);
+
+private:
+    class LayoutMapImpl;
+    LayoutMapImpl* impl;
+};
