+++ src/modules/keyboardmanager/common/KeyboardManagerCommon.vcxproj
@@ -90,7 +90,6 @@
     <ClCompile Include="Helpers.cpp" />
     <ClCompile Include="KeyboardManagerState.cpp" />
     <ClCompile Include="KeyDelay.cpp" />
-    <ClCompile Include="LayoutMap.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
@@ -103,7 +102,6 @@
     <ClInclude Include="KeyboardManagerConstants.h" />
     <ClInclude Include="KeyboardManagerState.h" />
     <ClInclude Include="KeyDelay.h" />
-    <ClInclude Include="LayoutMap.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="RemapShortcut.h" />
     <ClInclude Include="Shortcut.h" />
